[FE] rename: 폴더명 변경
