+++ frontend/src/components/public-space/best-project-box/wave-background/index.tsx
